consolidate python and utils tokenization files, they are copies
